contact form refactor
